Failed attempt to fix the sort feature
