+++ src/services/housing.js
@@ -170,51 +170,6 @@ const getApartmentApplication = async (applicationID) => {
  * @return {Promise.<ApplicationDetails>[]} Application details
  */
 const getAllApartmentApplications = async () => {
-  // let dummyApplicationDetails = {
-  //   ApplicationID: 42,
-  //   DateSubmitted: '2030-03-14',
-  //   DateModified: '2030-03-14',
-  //   EditorUsername: 'Bobby.Tables',
-  //   Gender: 'M',
-  //   Applicants: [
-  //     {
-  //       Username: 'Bobby.Tables',
-  //       Age: 21,
-  //       OffCampusProgram: 'Computer Science',
-  //       Probation: 'no',
-  //       Points: 7,
-  //     },
-  //     { Username: 'Frederick.Fox', Age: 20, OffCampusProgram: '', Probation: 'yes', Points: 5 },
-  //     {
-  //       Username: 'Tommy.Turtle',
-  //       Age: 22,
-  //       OffCampusProgram: 'Education',
-  //       Probation: 'no',
-  //       Points: 6,
-  //     },
-  //     {
-  //       Username: 'Tommy2.Turtle',
-  //       Age: 22,
-  //       OffCampusProgram: 'Education',
-  //       Probation: 'no',
-  //       Points: 6,
-  //     },
-  //     {
-  //       Username: 'Tommy3.Turtle',
-  //       Age: 22,
-  //       OffCampusProgram: 'Education',
-  //       Probation: 'no',
-  //       Points: 6,
-  //     },
-  //   ],
-  //   ApartmentChoices: [
-  //     { HallRank: 1, HallName: 'Gantley' },
-  //     { HallRank: 2, HallName: 'Tavilla' },
-  //   ],
-  // }; //! DEBUG: This exists purely for testing the features without the backend. The commented-out line below is the actual code to use once the endpoint has been created in the backend
-  // let { ...dummyData2 } = dummyApplicationDetails;
-  // dummyData2.ApplicationID = 43;
-  // let applicationDetailsArray = [dummyApplicationDetails, dummyData2];
   let applicationDetailsArray = await http.get(`housing/admin/apartment/applications/`);
 
   // Calculate the total and average points for each application
